(1) data tables changed from single quote to double quote for the string valued items to prevent post/put bugs; (2) contribution has modify function, but it's not populating fully and creates bugs for adding a new contributon under a donor specific view; (3) adding a new contribution entry is bugged and stores undefined for some reason; (4) contribution is now split up into multiple controllers to handle one for main view, specific view, and form entry view; (5) pledge is working fine, but needs to have multiple controllers like contributions AND needs modify function

+++ Node/M4K_Tables.sql
@@ -127,7 +127,7 @@ CREATE TABLE `Pledge` (
 `donor_id` int,
 `patient_id` int,
 `pledge_date` DATE,
-`target_amount` DECIMAL,
+`target_amount` DOUBLE,
 `is_behind` BOOL,
 PRIMARY KEY(pledge_id),
 FOREIGN KEY(donor_id) REFERENCES Donor(supporter_id) ON DELETE CASCADE,
@@ -137,7 +137,7 @@ FOREIGN KEY(patient_id) REFERENCES Patient(patient_id) ON DELETE CASCADE
 DROP TABLE IF EXISTS `Installments`;
 CREATE TABLE `Installments` (
 `pledge_id` int,
-`amount` DECIMAL,
+`amount` DOUBLE,
 `installment_date` DATE,
 FOREIGN KEY(pledge_id) REFERENCES Pledge(pledge_id) ON DELETE CASCADE
 );
